requirements.txt
annotated-types==0.7.0
numpy==2.0.1
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
